moves file
